requirements.txt
annotated-types==0.6.0
anyio==4.3.0
argcomplete==3.2.3
async-timeout==4.0.3
asyncpg==0.29.0
beautifulsoup4==4.12.3
bs4==0.0.2
build==1.1.1
CacheControl==0.14.0
certifi==2024.2.2
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
colorama==0.4.6
crashtest==0.4.1
distlib==0.3.8
dulwich==0.21.7
exceptiongroup==1.2.0
fastapi==0.110.0
fastjsonschema==2.19.1
filelock==3.13.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
importlib_metadata==7.0.2
installer==0.7.0
jaraco.classes==3.3.1
keyring==24.3.1
more-itertools==10.2.0
msgpack==1.0.8
packaging==24.0
pexpect==4.9.0
pillow==10.2.0
pipx==1.4.3
pkginfo==1.10.0
platformdirs==4.2.0
poetry==1.8.2
poetry-core==1.9.0
poetry-plugin-export==1.6.0
postgres==4.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
psycopg2-pool==1.2
ptyprocess==0.7.0
pydantic==2.6.4
pydantic_core==2.16.3
pyproject_hooks==1.0.0
python-dotenv==1.0.1
pywin32-ctypes==0.2.2
rapidfuzz==3.6.2
requests==2.31.0
requests-toolbelt==1.0.0
shellingham==1.5.4
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.28
starlette==0.36.3
tomli==2.0.1
tomlkit==0.12.4
trove-classifiers==2024.3.3
typing_extensions==4.10.0
urllib3==2.2.1
userpath==1.9.2
uvicorn==0.28.0
virtualenv==20.25.1
zipp==3.18.0
